Rebased over latest master and squashed previous commits:
eca9806013da49274d9b3f8b40c09002d8503f05 Use new function name for coinbase maturity; after rebase over 4.0.2
dec93ff3480b85dd01b56ea65bc191f28a06c670 Correct build config
3553a3e3f9d8cad19c7e3af029355e71b1fd6944 Move all zerocoin-related checks/functions to the consensus folder
8612a8ad6377a257910834896ec29f3fb4372903 Swap to newer method for genesis calculation; introduce Consensus namespace
3f19ccc5987a4acd4bfcd23070865828c50f4401 Missed includes and entry in CMakeLists.txt allowing tests to run
c6924f3ecbac8d075ebba80e2dfac8e17bdfbdf4 Move transaction checks out to consensus/tx_verify.cpp
2534662c944af71e200aa54f536b4312e4ba2a29 Move CDiskTxPos/CBlockUndo to txdb.h/undo.h respectively
2ab2e18c307631b04ecad5a490ee044068e0a5fd Move CValidationState class/messages to consensus folder
19d5af718667ecafed5a48127f85c82102099c9f Split some classes and const definitions out
