+++ src/test/zerocoin_coinspend_tests.cpp
@@ -12,6 +12,7 @@
 #include "amount.h"
 #include "chainparams.h"
 #include "coincontrol.h"
+#include "consensus/zerocoin_verify.h"
 #include "main.h"
 #include "wallet/wallet.h"
 #include "wallet/walletdb.h"
